fix(Environment): remove circular dependencies between packages
