+++ packages/core/src/lib/config/config.interface.ts
@@ -1,7 +1,9 @@
-import { IgoOptions } from '../environment/environment.interface';
+import { CoreOptions } from '../environment';
 
-export interface ConfigOptions {
-  default?: IgoOptions;
+export type BaseConfigOptions<T> = CoreOptions & T;
+
+export interface ConfigOptions<T = { [key: string]: any }> {
+  default?: BaseConfigOptions<T>;
   path?: string;
 }
