+++ packages/core/src/lib/environment/environment.interface.ts
@@ -1,68 +1,10 @@
-import { AuthOptions, AuthStorageOptions } from '@igo2/auth';
-import { ContextServiceOptions } from '@igo2/context';
-import {
-  CatalogServiceOptions,
-  CommonVectorStyleOptions,
-  ImportExportServiceOptions,
-  Projection,
-  SearchSourceOptions,
-  StoredQueriesReverseSearchSourceOptions,
-  StoredQueriesSearchSourceOptions
-} from '@igo2/geo';
 import { LanguageOptions } from '../language';
-import { DOMOptions } from '@igo2/common';
 
-export interface EnvironmentOptions {
+export interface BaseEnvironmentOptions {
   production: boolean;
-  igo: IgoOptions;
+  igo: CoreOptions;
 }
 
-export interface IgoOptions {
-  app?: AppOptions;
-  auth?: AuthOptions;
-  catalog?: CatalogServiceOptions;
-  context?: ContextServiceOptions;
-  depot?: { url: string; trainingGuides?: string[] };
-  dom?: DOMOptions[];
-  importExport?: ImportExportServiceOptions;
-  importWithStyle?: boolean;
-  interactiveTour?: InteractiveTourConfigOptions;
+export interface CoreOptions {
   language?: LanguageOptions;
-  projections?: Projection[];
-  queryOverlayStyle?: {
-    base?: CommonVectorStyleOptions;
-    selection?: CommonVectorStyleOptions;
-    focus?: CommonVectorStyleOptions;
-  };
-  searchOverlayStyle?: {
-    base?: CommonVectorStyleOptions;
-    selection?: CommonVectorStyleOptions;
-    focus?: CommonVectorStyleOptions;
-  };
-  searchSources?: {
-    [key: string]:
-      | SearchSourceOptions
-      | StoredQueriesSearchSourceOptions
-      | StoredQueriesReverseSearchSourceOptions;
-  };
-  storage?: AuthStorageOptions;
-}
-
-export interface AppOptions {
-  forceCoordsNA: boolean;
-  install: {
-    enabled?: boolean;
-    promote?: boolean;
-    manifestPath?: string;
-  };
-  pwa?: {
-    enabled?: boolean;
-    path?: string;
-  };
-}
-
-export interface InteractiveTourConfigOptions {
-  activateInteractiveTour: boolean;
-  tourInMobile: boolean;
-  pathToConfigFile?: string;
 }
